+++ src/main/java/movieApi/movies/dto/response/MovieDTO.java
@@ -1,4 +1,4 @@
-package movieApi.movies.dto;
+package movieApi.movies.dto.response;
 
 import java.util.List;
